fix migration file

+++ src/db/migrations/20200402151631_products.js
@@ -12,11 +12,11 @@ exports.up = function (knex) {
         table.text('image2').nullable()
         table.text('image3').nullable()
         table.text('description').notNullable()
-        table.float('quantity').defaultTo(0)
+        table.integer('quantity').defaultTo(0)
         table.boolean('availabilty').defaultTo(false)
         table.text('color').nullable()
-        table.integer('price').unsigned().notNullable()
-        table.integer('size').nullable()
+        table.float('price').unsigned().notNullable()
+        table.float('size').nullable()
         table.timestamp('created_at').defaultTo(knex.fn.now());
         table.timestamp('updated_at').nullable();
     })
